Add generic test-utils.cjs for JS tests
